test-web: Move convert_ansi_to_html to CaptureFile.cpp
