+++ Tests/LibWeb/test-web/Debug.h
@@ -6,14 +6,10 @@
 
 #pragma once
 
-#include <AK/StringBuilder.h>
-#include <AK/StringView.h>
 #include <AK/Types.h>
 
 namespace TestWeb {
 
 void maybe_attach_on_fail_fast_timeout(pid_t);
-
-StringBuilder convert_ansi_to_html(StringView);
 
 }
